+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## [Unreleased](https://github.com/tetsuo13/Roundcube-Matomo-Tracking-Api-Plugin/compare/2.0.0...main)
 
 - Renamed from Piwik to Matomo.
-- Removed bundled matomo-php-tracker in favor of using Composer.
+- Removed bundled matomo-php-tracker in favor of using Composer to handle dependencies.
 - Addressed many of the PHP deprecated warnings that were being logged ([#11](https://github.com/tetsuo13/Roundcube-Matomo-Tracking-Api-Plugin/issues/11))
+- Silently abort if server vars used to determine Roundcube installation URL are missing ([#11](https://github.com/tetsuo13/Roundcube-Matomo-Tracking-Api-Plugin/issues/11))
 
 ## [2.0.0](https://github.com/tetsuo13/Roundcube-Matomo-Tracking-Api-Plugin/compare/1.1.0...2.0.0) (2016-12-06)